refactor: extract, move and rename components from QuizContainerNew folder
