--- src/components/QuizContainerNew/QuestionsListNew/QuestionListItemNew/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./QuestionListItem";
-export type { QuestionListItemProps } from "./QuestionListItem";
