+++ src/components/TeacherComponents/QuestionListItemNew/QuizImage.test.tsx
