+++ src/components/TeacherComponents/QuestionListItemNew/QuizImageAnswer.test.tsx
