+++ src/components/TeacherComponents/QuestionListItemNew/QuizOakImage.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, FC, SetStateAction } from "react";
 
-import { calcDims } from "../../quizUtils";
+import { calcDims } from "../LessonOverviewQuizContainer/quizUtils";
 
 import OwaImage from "@/components/SharedComponents/OwaImage";
 import { StemImageObject } from "@/node-lib/curriculum-api-2023/shared.schema";
