+++ src/components/TeacherComponents/QuizQuestionsListItem/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./QuizQuestionsListItem";
+export type { QuizQuestionsListItemProps as QuestionListItemProps } from "./QuizQuestionsListItem";
